requirements.txt
flask>=3.0.0
requests>=2.31.0
flask-sock>=0.7.0
websocket-client>=1.6.0
